Rewrite 68k-z80 communications to prevent conflicts
When using a single flag variable, it's possible that 68k is writing to
it just before Z80 reads from it (in order to clear a bit). The bit that
was set by the 68k is then lost and you'll probably get a missed note or
wrong sample being played.

To fix, simply put each flag in a separate byte.